dom/base/test/jsmodules/test_moduleScriptsRun.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Test script elements with type="module" are run for chrome HTML</title>
<script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
<script>
  var inlineModuleRan = false;
  var moduleRan = false;

  SimpleTest.waitForExplicitFinish();

  // eslint-disable-next-line no-unused-vars
  function testLoaded() {
    ok(moduleRan, 'Check module script was run');
    ok(inlineModuleRan, 'Check inline module script was run');
    SimpleTest.finish();
  }
</script>
<script type="module" src="module_setRan.mjs"></script>
<script type="module">
  /* global inlineModuleRan */
  inlineModuleRan = true;
  </script>
<body onload='testLoaded()'></body>
